Finished implementing new webapp nginx upstream module.

+++ webapp/include/Database.h
@@ -33,7 +33,7 @@ class Database;
 class Query {
 public:
 	int status = DATABASE_QUERY_INIT;
-	long long lastrowid = 0;
+	uint64_t lastrowid = 0;
 	int column_count = 0;
 	webapp_str_t* row = NULL;
 	webapp_str_t* description = NULL;
@@ -70,7 +70,7 @@ public:
 
 	int connect(int database_type, const char* host, const char* username, 
 		const char* password, const char* database);
-	long long exec(const std::string& query);
+	uint64_t exec(const std::string& query);
 };
 
 #endif //DATABASE_H
